+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- Show error if there is invalid column in "EXCEPT" clause of "SELECT" command.
+- Minor improvements and optimizations.
+
 ## [0.7.0] - 2024-09-12
 
 ### Added